glm pojo: ignore unseen levels when told to do so
PUBDEV-3543

Note that this fix has a side effect of also ignoring the case in which
a given parameter isn't provided. I'm assuming this is fine and a more
involved fix would be required to differentiate between missing
parameters and unseen categorical levels (as they both enter `score0` as
NaN).